fix(build): postinstall, only remove kit bin if windows

+++ build/postinstall.js
@@ -13,18 +13,18 @@ if (isCI) {
 			...pathParts
 		)
 
-	console.log(`Run postinstall on ${kitPath()}`)
-
-	let binFileToRemove = platform() === "win32" ? "kit" : "kit.bat"
-	let binFilePathToRemove = kitPath("bin", binFileToRemove)
-	console.log(`Checking if ${binFilePathToRemove} exists...`)
-	
-	if (existsSync(binFilePathToRemove)) {
-		console.log(
-			`Removing ${kitPath("bin", binFileToRemove)} so it doesn't interfere with kit.bat`
-		)
-		await rimraf(kitPath("bin", binFileToRemove))
-	} else {
-		console.log(`${binFilePathToRemove} does not exist. Skipping...`)
+	if (platform() === "win32") {
+		let binFileToRemove = "kit"
+		let binFilePathToRemove = kitPath("bin", binFileToRemove)
+		console.log(`Checking if ${binFilePathToRemove} exists...`)
+		
+		if (existsSync(binFilePathToRemove)) {
+			console.log(
+				`Removing ${kitPath("bin", binFileToRemove)} so it doesn't interfere with kit.bat`
+			)
+			await rimraf(kitPath("bin", binFileToRemove))
+		} else {
+			console.log(`${binFilePathToRemove} does not exist. Skipping...`)
+		}
 	}
 }
